tools/bpf: add support to read .BTF.ext sections
To support BTF, an .BTF section has been added to elf file
which encodes types. These types can be used for map
pretty print. The whole .BTF will be passed to the
kernel as well for which kernel can verify and return
to the user space for pretty print etc.

An .BTF.ext has been proposed to be added to elf file
which will include data which are useful for
userspace only. Right now, it will only contain func
type info which userspace will do necessary preprocessing
and only pass needed information to kernel through
bpf_prog_load syscall interface.

This functionality is implemented in libbpf.
The header for .BTF.ext is very similar to .BTF section
so it can be extended in the future.

Right now, the .BTF.ext only contain func_info secton.
This infomation will be used to contruct func type info
specific for each loadable program, which will then be
passed to the kernel through bpf_prog_load syscall
interface.

Signed-off-by: Yonghong Song <yhs@fb.com>