tools/bpf: add support to read .BTF.ext sections
Previously, an .BTF section has been added to elf file
which encodes types. These types can be used for map
pretty print. The whole .BTF will be passed to the
kernel as well for which kernel can verify and return
to the user space for pretty print etc.

An .BTF.ext has been proposed to be added to elf file
which will include func type info and debug line info.
In this patch, only func type info is processed.
A separate section is used since the bpf loader
needs to manipulate the information before
passing to the kernel, if the bytecodes are scattered
in more than one sections.

This functionality is implemented in libbpf.
The header for .BTF.ext is very similar to .BTF section
so it can be extended in the future.

Signed-off-by: Yonghong Song <yhs@fb.com>